requirements.txt
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==45.0.5
cycler==0.12.1
distlib==0.3.8
dj-rest-auth==7.0.1
Django==5.1.2
django-allauth==65.8.0
django-cors-headers==4.7.0
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.6.1
filelock==3.16.1
fonttools==4.51.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.40.3
h11==0.14.0
html5lib==1.1
huffman==0.1.2
idna==3.10
imageio==2.34.2
joblib==1.4.2
kaggle==1.6.17
kiwisolver==1.4.5
matplotlib==3.8.4
mlfz==0.1.3.2
mongopy==0.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.2.4
networkx==3.3
nltk==3.8.1
nose==1.3.7
numpy==1.26.4
oauthlib==3.3.1
opendatasets==0.1.22
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
pillow==10.3.0
platformdirs==4.3.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.21.0
pygame==2.5.2
PyJWT==2.9.0
pymongo==4.6.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
regex==2024.5.15
requests==2.32.4
rsa==4.9.1
scipy==1.14.1
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
tk==0.1.0
tqdm==4.66.4
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
